kong/lua: fix broken kong deployment
Kong does no longer generate its start script in the bin directory of
the lua runtime when installed via luarocks. So instead we install from
source, which for the most part does the exact same thing
Meanwhile, according to the kong docs
(https://getkong.org/install/source/), it is recommended to use the lua
runtime that comes with openresty.